+++ src/other/index.jl
@@ -229,8 +229,20 @@ end
 @inline Base.getindex(x::AbstractIndex, idx::Between) = x[idx.first]:x[idx.last]
 @inline Base.getindex(x::AbstractIndex, idx::All) =
     isempty(idx.cols) ? (1:length(x)) : throw(ArgumentError("All(args...) is not supported: use Cols(args...) instead"))
-@inline Base.getindex(x::AbstractIndex, idx::Cols) =
-    isempty(idx.cols) ? Int[] : union(getindex.(Ref(x), idx.cols)...)
+
+@inline function Base.getindex(x::AbstractIndex, idx::Cols) =
+    isempty(idx.cols) && return Int[]
+    if idx.operation == :union
+        return union(getindex.(Ref(x), idx.cols)...)
+    elseif idx.operation == :intersect
+        return intersect(getindex.(Ref(x), idx.cols)...)
+    else
+        throw(ArgumentError("Unsupported operation :$(idx.operation). Either " *
+                            ":union or :intersect is required"))
+    end
+end
+# the definition below is needed because `:` is Function
+
 @inline Base.getindex(x::AbstractIndex, idx::Cols{Tuple{typeof(:)}}) = x[:]
 @inline Base.getindex(x::AbstractIndex, idx::Cols{<:Tuple{Function}}) =
     findall(idx.cols[1], names(x))